Ported replace into tests
